Move out metadata and end user providers
